// src/app/shared/models/questionnaire.model.ts
export interface Questionnaire {
   adaptestSettingsFieldId: string;
   allowEdit: false;
   displayName: string;
   displaySummaryScore: false;
   enableBusinessRuleId: string;
   enabledBusinessRuleResult: boolean;
   enableVisitQuestionnaireId: string;
   enforcePreviousDaysEntry: false;
   id: string;
   internalName: string;
   isBlinded: false;
   isDropdownNavigationEnabled: false;
   openEndedBusinessRuleId: string;
   previousDaysEdit: false;
   previousDaysEntry: number;
   promptIncompletePagesOnSave: false;
   questionnaireTypeId: number;
   sequence: number;
   translationKey: string;
   url: string;
   validateQuestionnaireOnSave: false;
   visibleBusinessRuleId: string;
   visibleBusinessRuleResult: boolean;
}